Add files via upload
Fixed numerous issues discussed on Discord -- https://discord.com/channels/499582383067234305/1024403618784362506/1156200984155004988

"First, the curly bracket {tags} are actually correct in this case, for this particular file (some files have specific rules for their tags, which seem to change from file to file! This *also* applies below... )

The {tags} here are {Case_Sensitive} and in this case, every tag needs to be {Upper_Case}. In this instance, all the {PC} tags, which were preserved from the original file, were changed into {pc} to match every other file, which caused this issue. The tag needed to be {PC} with this particular file

It *also* appears that the {PC} tag was added to certain lines which did not have this tag originally. Only lines that expect {PC} will actually properly replace this tag! Otherwise, it will literally just spit out {PC}. So now i'm going back and retranslating these lines."